Performed EDA, graph ploting, Decision Tree model
Heart Disease Pacients - Exploratory Data Analysis and Prediction (Python Implementation)

1. The dataset contain records of pacients with or without heart diseases (and other sickness).
2. Performed Data Exploration, Data Wrangling and Visualization.
3. Model used: Decision Tree.
4. Libs used: Pandas, Numpy, Sklearn, Seaborn, Graphviz

Accuracy = 0.8640803010246094

Disclaimer: Most of the hits of the model are the TN(True Negatives). I imagine that happens because 91% of the observations are "NO" for the predictable varible "HeartDisease".
This could be minimized if we subset into a more fair distribution between "NO" and "YES" for the predict variable besides other adjustments into the tree.